test: Move test project to test folder and read test data as embedded resources
