+++ AppLibDotnet.sln
@@ -17,7 +17,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{6C8EB054-1
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{7AD5FADE-607F-4D5F-8511-6647D0C1AA1C}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Altinn.App.Common.Tests", "Altinn.App.Common.Tests\Altinn.App.Common.Tests.csproj", "{95B934FA-4EA2-4893-9F4B-D98384ACE74C}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Altinn.App.Common.Tests", "test\Altinn.App.Common.Tests\Altinn.App.Common.Tests.csproj", "{CF47083E-B3E0-4247-87AA-F7EF6B3292D1}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -45,10 +45,10 @@ Global
 		{2FD56505-1DB2-4AE1-8911-E076E535EAC6}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{2FD56505-1DB2-4AE1-8911-E076E535EAC6}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{2FD56505-1DB2-4AE1-8911-E076E535EAC6}.Release|Any CPU.Build.0 = Release|Any CPU
-		{95B934FA-4EA2-4893-9F4B-D98384ACE74C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{95B934FA-4EA2-4893-9F4B-D98384ACE74C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{95B934FA-4EA2-4893-9F4B-D98384ACE74C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{95B934FA-4EA2-4893-9F4B-D98384ACE74C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{CF47083E-B3E0-4247-87AA-F7EF6B3292D1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CF47083E-B3E0-4247-87AA-F7EF6B3292D1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CF47083E-B3E0-4247-87AA-F7EF6B3292D1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CF47083E-B3E0-4247-87AA-F7EF6B3292D1}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -59,7 +59,6 @@ Global
 		{98E6200A-ED99-418E-B30C-81BA564B509A} = {7AD5FADE-607F-4D5F-8511-6647D0C1AA1C}
 		{17D7DCE9-7797-4BC1-B448-D0529FD6FB3D} = {6C8EB054-1747-4BAC-A637-754F304BCAFA}
 		{2FD56505-1DB2-4AE1-8911-E076E535EAC6} = {6C8EB054-1747-4BAC-A637-754F304BCAFA}
-		{95B934FA-4EA2-4893-9F4B-D98384ACE74C} = {6C8EB054-1747-4BAC-A637-754F304BCAFA}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {4584C6E1-D5B4-40B1-A8C4-CF4620EB0896}
