+++ test/Altinn.App.Common.Tests/Altinn.App.Common.Tests.csproj
@@ -8,6 +8,16 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <None Remove="Helpers\JsonHelperTests\TestData\simple_changed.json" />
+    <None Remove="Helpers\JsonHelperTests\TestData\simple_unchanged.json" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <EmbeddedResource Include="Helpers\JsonHelperTests\TestData\simple_changed.json" />
+    <EmbeddedResource Include="Helpers\JsonHelperTests\TestData\simple_unchanged.json" />
+  </ItemGroup>
+
+  <ItemGroup>
     <PackageReference Include="FluentAssertions" Version="6.7.0" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.11.0" />
     <PackageReference Include="xunit" Version="2.4.1" />
@@ -22,7 +32,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\src\Altinn.App.Common\Altinn.App.Common.csproj" />
+    <ProjectReference Include="..\..\src\Altinn.App.Common\Altinn.App.Common.csproj" />
   </ItemGroup>
 
 </Project>
