+++ test/Altinn.App.Common.Tests/Helpers/JsonHelperTests/TestData/simple_changed.json
@@ -1,0 +1,3 @@
+{
+  "test" : "to_be_removed"
+}
