+++ test/Altinn.App.PlatformServices.Tests/Altinn.App.PlatformServices.Tests.csproj
@@ -39,6 +39,10 @@
     </AdditionalFiles>
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Helpers\NewFolder\" />
+  </ItemGroup>
+
   <PropertyGroup Condition="'$(Configuration)'=='Debug'">
     <CodeAnalysisRuleSet>..\..\Altinn3.ruleset</CodeAnalysisRuleSet>
   </PropertyGroup>
